Move format-data to store directory
