+++ src/store/load-data.js
@@ -1,7 +1,7 @@
 import { json } from 'd3-fetch';
 import config from '../config';
 import getRandomPipeline from '../utils/random-data';
-import formatData from '../utils/format-data';
+import formatData from './format-data';
 import loremIpsum from '../utils/data/lorem-ipsum.mock';
 import animals from '../utils/data/animals.mock';
 import demo from '../utils/data/demo.mock';
