+++ src/store/load-data.test.js
@@ -1,5 +1,5 @@
 import loadData from './load-data';
-import formatData from '../utils/format-data';
+import formatData from './format-data';
 import loremIpsum from '../utils/data/lorem-ipsum.mock';
 import animals from '../utils/data/animals.mock';
 import demo from '../utils/data/demo.mock';
